+++ tests/services/detectionlists/test_high_risk_employee.py
@@ -1,6 +1,5 @@
 import pytest
-from requests import HTTPError
-from requests import Response
+from tests.conftest import create_mock_error
 
 from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42NotFoundError
@@ -77,10 +76,9 @@ class TestHighRiskEmployeeClient(object):
     ):
         def side_effect(url, json):
             if "add" in url:
-                base_err = mocker.MagicMock(spec=HTTPError)
-                base_err.response = mocker.MagicMock(spec=Response)
-                base_err.response.text = "User already on list"
-                raise Py42BadRequestError(base_err)
+                raise create_mock_error(
+                    Py42BadRequestError, mocker, "User already on list"
+                )
 
         mock_connection.post.side_effect = side_effect
         client = HighRiskEmployeeService(