+++ tests/services/detectionlists/test_user_profile.py
@@ -1,12 +1,10 @@
 import pytest
-from requests import Response
-from requests.exceptions import HTTPError
+from tests.conftest import create_mock_error
 
 from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42CloudAliasLimitExceededError
 from py42.services.detectionlists.user_profile import DetectionListUserService
 from py42.services.users import UserService
-
 
 CLOUD_ALIAS_LIMIT_EXCEEDED_ERROR_MESSAGE = """{
 "pop-bulletin": {
@@ -28,11 +26,9 @@ class TestDetectionListUserClient(object):
 
     @pytest.fixture
     def mock_get_by_id_fails(self, mocker, mock_connection):
-        response = mocker.MagicMock(spec=Response)
-        response.status_code = 400
-        exception = mocker.MagicMock(spec=HTTPError)
-        exception.response = response
-        mock_connection.post.side_effect = Py42BadRequestError(exception)
+        mock_connection.post.side_effect = create_mock_error(
+            Py42BadRequestError, mocker, ""
+        )
         return mock_connection
 
     @pytest.fixture
@@ -40,11 +36,9 @@ class TestDetectionListUserClient(object):
         self, mocker, mock_connection, user_context, py42_response
     ):
         user_client = UserService(mock_connection)
-        response = mocker.MagicMock(spec=Response)
-        response.status_code = 400
-        exception = mocker.MagicMock(spec=HTTPError)
-        exception.response = response
-        mock_connection.post.side_effect = Py42BadRequestError(exception)
+        mock_connection.post.side_effect = create_mock_error(
+            Py42BadRequestError, mocker, ""
+        )
         return user_client
 
     @pytest.fixture
@@ -52,32 +46,10 @@ class TestDetectionListUserClient(object):
         self, mocker, mock_connection, user_context, py42_response
     ):
         user_client = UserService(mock_connection)
-        response = mocker.MagicMock(spec=Response)
-        response.status_code = 400
-        response.text = CLOUD_ALIAS_LIMIT_EXCEEDED_ERROR_MESSAGE
-        exception = mocker.MagicMock(spec=HTTPError)
-        exception.response = response
-        mock_connection.post.side_effect = Py42BadRequestError(exception)
+        mock_connection.post.side_effect = create_mock_error(
+            Py42BadRequestError, mocker, CLOUD_ALIAS_LIMIT_EXCEEDED_ERROR_MESSAGE
+        )
         return user_client
-
-    def test_create_posts_expected_data(
-        self, mock_connection, user_context, mock_user_client
-    ):
-        detection_list_user_client = DetectionListUserService(
-            mock_connection, user_context, mock_user_client
-        )
-        detection_list_user_client.create("942897397520289999")
-
-        posted_data = mock_connection.post.call_args[1]["json"]
-        assert mock_connection.post.call_count == 1
-        assert mock_connection.post.call_args[0][0] == "v2/user/create"
-        assert (
-            posted_data["tenantId"] == user_context.get_current_tenant_id()
-            and posted_data["userName"] == "942897397520289999"
-            and posted_data["riskFactors"] == []
-            and posted_data["cloudUsernames"] == []
-            and posted_data["notes"] == ""
-        )
 
     def test_get_posts_expected_data(
         self, mock_connection, user_context, mock_user_client